requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.6.1
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0
